requirements.txt

asgiref3.6.0

Django4.1.7

mysqlclient2.1.1

pip23.0.1
setuptools  65.5.0
sqlparse    0.4.3
